Test configurations of methods based on base config

+++ src/main/java/com/pturpin/quickcheck/test/configuration/TestRunnerConfiguration.java
@@ -5,7 +5,7 @@ package com.pturpin.quickcheck.test.configuration;
  */
 public interface TestRunnerConfiguration {
 
-  int getNbRun();
+  long getNbRun();
 
   boolean acceptSkipped();
